requirements.txt
absl-py==0.12.0
asgiref==3.3.1
astunparse==1.6.3
atomicwrites==1.4.0
attrs==21.4.0
beautifulsoup4==4.9.3
bs4==0.0.1
cached-property==1.5.2
cachetools==4.2.2
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
colorama==0.4.5
coreapi==2.3.3
coreschema==0.0.4
coverage==6.4.1
cycler==0.10.0
DateTime==4.3
distro==1.6.0
Django==3.1.6
django-crispy-forms==1.14.0
djangorestframework==3.13.1
djangorestframework-api-key==2.2.0
drf-yasg==1.20.0
feedparser==6.0.2
finnhub-python==2.4.11
finvizfinance==0.13.0
gast==0.3.3
greenlet==1.1.2
grpcio==1.37.1
html5lib==1.1
idna==2.10
importlib-metadata==4.0.1
inflection==0.5.1
iniconfig==1.1.1
itsdangerous==2.0.1
itypes==1.2.0
Jinja2==3.1.1
joblib==1.0.1
kiwisolver==1.3.1
lxml==4.6.2
Markdown==3.3.4
MarkupSafe==2.1.1
matplotlib==3.4.2
multitasking==0.0.9
mysql-connector-python==8.0.28
mysqlclient==2.1.0
nltk==3.5
numpy==1.19.5
oauthlib==3.1.0
openapi-codec==1.3.2
opt-einsum==3.3.0
packaging==21.3
pandas==1.2.2
Pillow==8.2.0
plotly==5.3.1
pluggy==1.0.0
pprintpp==0.4.0
praw==7.2.0
prawcore==2.0.0
protobuf==3.16.0
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycoingecko==2.2.0
pycountry==22.3.5
pycountry-convert==0.7.2
pyparsing==2.4.7
PySocks==1.7.1
pytest==7.1.2
pytest-cov==3.0.0
pytest-mock==3.8.1
python-dateutil==2.8.1
pytrends==4.7.3
pytz==2021.1
PyYAML==5.4.1
regex==2020.11.13
repoze.lru==0.7
requests==2.25.1
requests-cache==0.6.4
requests-oauthlib==1.3.0
rsa==4.7.2
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
scikit-learn==0.24.2
scipy==1.7.3
sgmllib3k==1.0.0
simplejson==3.17.6
six==1.15.0
soupsieve==2.2.1
SQLAlchemy==1.4.32
sqlparse==0.4.1
tabula-py==2.3.0
tenacity==8.0.1
termcolor==1.1.0
threadpoolctl==2.1.0
tomli==2.0.1
tqdm==4.56.2
tweepy==3.10.0
typing-extensions==3.10.0.0
update-checker==0.18.0
uritemplate==4.1.1
url-normalize==1.4.3
urllib3==1.26.3
vaderSentiment==3.3.2
webencodings==0.5.1
websocket-client==0.58.0
Werkzeug==1.0.1
whitenoise==5.2.0
wrapt==1.12.1
xlrd==2.0.1
yfinance==0.1.63
zipp==3.4.1
zope.interface==5.4.0
